Rename app package to "fi.fenhua.android.badminton.scoreboard".
