+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "22.0.1"
 
     defaultConfig {
-        applicationId "com.example.android.badmintonscoreboard"
+        applicationId "fi.fenhua.android.badminton.scoreboard"
         minSdkVersion 15
         targetSdkVersion 22
         versionCode 1
