+++ app/src/main/java/fi/fenhua/android/badminton/scoreboard/Game.java
@@ -1,4 +1,4 @@
-package com.example.android.badmintonscoreboard;
+package fi.fenhua.android.badminton.scoreboard;
 
 import java.util.List;
