+++ app/src/main/java/fi/fenhua/android/badminton/scoreboard/Utils.java
@@ -1,4 +1,4 @@
-package com.example.android.badmintonscoreboard;
+package fi.fenhua.android.badminton.scoreboard;
 
 import android.content.Context;
 import android.content.res.TypedArray;
