+++ app/src/main/res/layout/activity_main.xml
@@ -28,7 +28,7 @@
                 app:expandedTitleMarginStart="105dp"
                 app:layout_scrollFlags="scroll|exitUntilCollapsed">
 
-                <com.example.android.badmintonscoreboard.SquareImageView
+                <fi.fenhua.android.badminton.scoreboard.SquareImageView
                     android:id="@+id/homeImage"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
